cc/pyproject.toml
[tool.poetry]
name = "ecodo"
version = "0.1.0"
description = ""
authors = [""]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.24.0"
numpy = "^1.26.2"
tensorflow-io-gcs-filesystem = { version = "*", python = ">=3.10,<3.12" }
tensorflow = "^2.15.0"
pillow = "^9.5.0, <10.0.0"
cleo = "^2.1.0"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
